Gaseous now runs without error on MariaDB

+++ gaseous-server/Classes/Metadata/Storage.cs
@@ -428,7 +428,7 @@ namespace gaseous_server.Classes.Metadata
             Database db = new gaseous_tools.Database(Database.databaseType.MySql, Config.DatabaseConfiguration.ConnectionString);
             string sql = "SELECT EXISTS(SELECT * FROM information_schema.tables WHERE table_schema = '" + Config.DatabaseConfiguration.DatabaseName + "' AND table_name = '" + TableName + "');";
             DataTable data = db.ExecuteCMD(sql);
-            if ((long)data.Rows[0][0] == 0)
+            if ((int)data.Rows[0][0] == 0)
             {
                 // table doesn't exist, create it
                 sql = "CREATE TABLE `" + Config.DatabaseConfiguration.DatabaseName + "`.`" + TableName + "` (`" + PrimaryTable + "Id` BIGINT NOT NULL, `" + SecondaryTable + "Id` BIGINT NOT NULL, PRIMARY KEY (`" + PrimaryTable + "Id`, `" + SecondaryTable + "Id`), INDEX `idx_PrimaryColumn` (`" + PrimaryTable + "Id` ASC) VISIBLE);";
